+++ src/main/java/pique/runnable/AQualityModelDeriver.java
@@ -33,6 +33,7 @@ import pique.model.QualityModel;
 import pique.model.QualityModelImport;
 import pique.utility.PiqueProperties;
 
+import java.math.BigDecimal;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Map;
@@ -53,7 +54,7 @@ public abstract class AQualityModelDeriver {
 
         // (1) Derive thresholds
         IBenchmarker benchmarker = qmDesign.getBenchmarker();
-        Map<String, Double[]> measureNameThresholdMappings = benchmarker.deriveThresholds(
+        Map<String, BigDecimal[]> measureNameThresholdMappings = benchmarker.deriveThresholds(
             benchmarkRepository, qmDesign, tools, projectRootFlag);
 
         // (2) Elicitate weights
